🚜: Restructure project files
