+++ src/main/java/com/example/superproject1/repository/item/Item.java
@@ -1,5 +1,8 @@
-package com.example.superproject1.repository.entity;
+package com.example.superproject1.repository.item;
 
+import com.example.superproject1.repository.entity.CartItem;
+import com.example.superproject1.repository.payment.PaymentItem;
+import com.example.superproject1.repository.entity.Sale;
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
