+++ src/main/java/com/example/superproject1/repository/item/ItemRepository.java
@@ -1,4 +1,4 @@
 package com.example.superproject1.repository.item;
 
-public class ItemRepository {
+public interface ItemRepository {
 }
