+++ src/main/java/com/example/superproject1/repository/users/User.java
@@ -1,7 +1,7 @@
 package com.example.superproject1.repository.users;
 
 import com.example.superproject1.repository.entity.Cart;
-import com.example.superproject1.repository.entity.Payment;
+import com.example.superproject1.repository.payment.Payment;
 import com.example.superproject1.repository.entity.Sale;
 import com.example.superproject1.repository.users.userRoles.UserRoles;
 import jakarta.persistence.*;
